requirements.txt
WooCommerce==3.0.0
cloudinary==1.41.0
python-dotenv~=0.15.0
ruamel.yaml==0.18.6
